+++ src/FormField.ts
@@ -6,12 +6,15 @@ import type { StandardSchemaV1 } from "@standard-schema/spec";
 
 export function newFormField<Data>(path: FieldPath, formAccess: FormAccess): FormField<Data> {
     const field: FormField<Data> = Object.assign(
-        (<K extends keyof Data>(pathKey: K): FormField<Data[K]> => {
-            if (typeof pathKey === "string" || typeof pathKey === "number") {
-                return newFormField(path.withProperty(pathKey), formAccess);
-            }
-            throw new Error("Unsupported path key " + pathKey.toString());
-        }),
+        Object.defineProperty(
+            (<K extends keyof Data>(pathKey: K): FormField<Data[K]> => {
+                if (typeof pathKey === "string" || typeof pathKey === "number") {
+                    return newFormField(path.withProperty(pathKey), formAccess);
+                }
+                throw new Error("Unsupported path key " + pathKey.toString());
+            }),
+            "name", { value: path.name() }
+        ),
         {
             toString: () => path.toString(),
             getData: () => formAccess.getData(path) as Data,
@@ -36,7 +39,7 @@ export function newFormField<Data>(path: FieldPath, formAccess: FormAccess): For
             _internal: {
                 addDeepErrorsListener: (subscriber: Subscriber) => formAccess.addDeepErrorsListener(path, subscriber),
             }
-        } satisfies BaseField<Data, true>, {
+        } satisfies Omit<BaseField<Data, true>, "name">, {
             push: (...element: any) => {
                 formAccess.setData(path, (data: unknown[]) => {
                     const copy = [...data];
@@ -60,6 +63,13 @@ export function newFormField<Data>(path: FieldPath, formAccess: FormAccess): For
 }
 
 type BaseField<Data, Writable extends boolean> = {
+    /**
+     * The field's name relative to its parent.
+     *
+     * For example, a field with the path `user.username` has the name `username`.
+     */
+    name: string
+
     /** Get the path of the field, joined with periods, e.g. "users.0.username" */
     toString: () => string