+++ CHANGELOG.md
@@ -2,6 +2,10 @@ Formula uses [semantic versioning](https://semver.org/).
 
 Formula is currently pre-v1. It's relatively stable, but I expect there will be some breaking changes before v1.
 
+# Unrelease
+
+- feat: `FormField.name` property
+
 # 0.0.14
 
 - fix: preserve non-standard StandardSchema.Issue fields